+++ src/SadConsole/Themes/ControlsConsoleTheme.cs
@@ -15,17 +15,6 @@
         public Cell FillStyle;
 
         /// <summary>
-        /// Creates a new controls console theme with the specified colors.
-        /// </summary>
-        /// <param name="themeColors">The colors used with this theme.</param>
-        public ControlsConsoleTheme(Colors themeColors) => RefreshTheme(themeColors);
-
-        /// <summary>
-        /// Creates a new theme without specifying the colors.
-        /// </summary>
-        protected ControlsConsoleTheme() { }
-
-        /// <summary>
         /// Returns a clone of this object.
         /// </summary>
         /// <returns>The cloned object.</returns>
@@ -38,6 +27,11 @@
             return newItem;
         }
 
+        public ControlsConsoleTheme()
+        {
+            FillStyle = Library.Default.Colors.Appearance_ControlNormal;
+        }
+
         /// <summary>
         /// Draws the theme to the console.
         /// </summary>
@@ -45,15 +39,14 @@
         /// <param name="hostSurface">Surface used for drawing.</param>
         public virtual void Draw(ControlsConsole console, CellSurface hostSurface)
         {
+            Colors colors = console.ThemeColors ?? Library.Default.Colors;
+
+            FillStyle = colors.Appearance_ControlNormal;
+
             hostSurface.DefaultForeground = FillStyle.Foreground;
             hostSurface.DefaultBackground = FillStyle.Background;
             hostSurface.Fill(hostSurface.DefaultForeground, hostSurface.DefaultBackground, FillStyle.Glyph, null);
         }
 
-        /// <summary>
-        /// Updates the theme with a color scheme.
-        /// </summary>
-        /// <param name="themeColors">The colors to update with.</param>
-        public virtual void RefreshTheme(Colors themeColors) => FillStyle = new Cell(themeColors.ControlHostFore, themeColors.ControlHostBack);
     }
 }