+++ src/SadConsole/Themes/Library.cs
@@ -1,5 +1,7 @@
 ﻿namespace SadConsole.Themes
 {
+    using System;
+    using System.Collections.Generic;
     using System.Runtime.Serialization;
     using SadConsole.Controls;
 
@@ -9,12 +11,36 @@
     [DataContract]
     public class Library
     {
+        private static Library _libraryInstance;
+
         private Colors _colors;
+        private Dictionary<System.Type, ThemeBase> _controlThemes;
 
         /// <summary>
         /// If a control does not specify its own theme, the theme from this property will be used.
         /// </summary>
-        public static Library Default { get; set; }
+        public static Library Default
+        {
+            get
+            {
+                if (_libraryInstance != null) return _libraryInstance;
+
+                _libraryInstance = new Library();
+                _libraryInstance.Init();
+
+                return _libraryInstance;
+            }
+            set
+            {
+                if (null == value)
+                {
+                    _libraryInstance = new Library();
+                    _libraryInstance.Init();
+                }
+                else
+                    _libraryInstance = value;
+            }
+        }
 
         /// <summary>
         /// Colors for the theme library.
@@ -25,76 +51,11 @@
             get => _colors;
             set
             {
-                if (_colors == null)
-                {
-                    throw new System.NullReferenceException("Colors cannot be set to null");
-                }
+                if (_colors == null) throw new System.NullReferenceException("Colors cannot be set to null");
 
                 _colors = value;
-
-                OnColorsChanged();
             }
         }
-
-        /// <summary>
-        /// Theme for the <see cref="Button"/> control.
-        /// </summary>
-        [DataMember]
-        public ButtonTheme ButtonTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="SelectionButton"/> control.
-        /// </summary>
-        [DataMember]
-        public ButtonTheme SelectionButtonTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="ScrollBar"/> control.
-        /// </summary>
-        [DataMember]
-        public ScrollBarTheme ScrollBarTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="RadioButton"/> control.
-        /// </summary>
-        [DataMember]
-        public RadioButtonTheme RadioButtonTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="ListBox"/> control.
-        /// </summary>
-        [DataMember]
-        public ListBoxTheme ListBoxTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="CheckBox"/> control.
-        /// </summary>
-        [DataMember]
-        public CheckBoxTheme CheckBoxTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="TextBox"/> control.
-        /// </summary>
-        [DataMember]
-        public TextBoxTheme TextBoxTheme { get; set; }
-
-        /// <summary>
-        /// Theme for the <see cref="ProgressBar"/> control.
-        /// </summary>
-        [DataMember]
-        public ProgressBarTheme ProgressBarTheme { get; set; }
-
-        /// <summary>
-        /// Theme for <see cref="DrawingSurface"/>.
-        /// </summary>
-        [DataMember]
-        public DrawingSurfaceTheme DrawingSurfaceTheme { get; set; }
-
-        /// <summary>
-        /// Theme for <see cref="Label"/>.
-        /// </summary>
-        [DataMember]
-        public LabelTheme LabelTheme { get; set; }
 
         /// <summary>
         /// Theme for <see cref="ControlsConsole"/>.
@@ -117,21 +78,24 @@
             }
         }
 
-        private void Init()
+        /// <summary>
+        /// Seeds the library with the default themes.
+        /// </summary>
+        protected void Init()
         {
-            ControlsConsoleTheme = new ControlsConsoleTheme(Colors);
-            WindowTheme = new WindowTheme(Colors);
+            ControlsConsoleTheme = new ControlsConsoleTheme();
+            WindowTheme = new WindowTheme();
 
-            ScrollBarTheme = new ScrollBarTheme();
-            ButtonTheme = new ButtonTheme();
-            CheckBoxTheme = new CheckBoxTheme();
-            ListBoxTheme = new ListBoxTheme(new ScrollBarTheme());
-            ProgressBarTheme = new ProgressBarTheme();
-            RadioButtonTheme = new RadioButtonTheme();
-            TextBoxTheme = new TextBoxTheme();
-            SelectionButtonTheme = new ButtonTheme();
-            DrawingSurfaceTheme = new DrawingSurfaceTheme();
-            LabelTheme = new LabelTheme();
+            SetControlTheme(typeof(ScrollBar), new ScrollBarTheme());
+            SetControlTheme(typeof(SelectionButton), new ButtonTheme());
+            SetControlTheme(typeof(Button), new ButtonTheme());
+            SetControlTheme(typeof(CheckBox), new CheckBoxTheme());
+            SetControlTheme(typeof(ListBox), new ListBoxTheme(new ScrollBarTheme()));
+            SetControlTheme(typeof(ProgressBar), new ProgressBarTheme());
+            SetControlTheme(typeof(RadioButton), new RadioButtonTheme());
+            SetControlTheme(typeof(TextBox), new TextBoxTheme());
+            SetControlTheme(typeof(DrawingSurface), new DrawingSurfaceTheme());
+            SetControlTheme(typeof(Label), new LabelTheme());
         }
 
         /// <summary>
@@ -140,7 +104,7 @@
         public Library()
         {
             _colors = new Colors();
-            Init();
+            _controlThemes = new Dictionary<Type, ThemeBase>(15);
         }
 
         /// <summary>
@@ -150,90 +114,44 @@
         private Library(bool _) => _colors = new Colors();
 
         /// <summary>
-        /// Gets a new control theme based on the control passed.
+        /// Creates and returns a theme based on the type of control provided.
         /// </summary>
         /// <param name="control">The control instance</param>
         /// <returns>A theme that is associated with the control.</returns>
-        public virtual ThemeBase GetControlTheme(ControlBase control)
+        public ThemeBase GetControlTheme(Type control)
         {
-            switch (control)
-            {
-                case SelectionButton c:
-                    return SelectionButtonTheme.Clone();
+            if (_controlThemes.ContainsKey(control))
+                return _controlThemes[control].Clone();
 
-                case ScrollBar c:
-                    return ScrollBarTheme.Clone();
-
-                case RadioButton c:
-                    return RadioButtonTheme.Clone();
-
-                case ListBox c:
-                    return ListBoxTheme.Clone();
-
-                case CheckBox c:
-                    return CheckBoxTheme.Clone();
-
-                case TextBox c:
-                    return TextBoxTheme.Clone();
-
-                case ProgressBar c:
-                    return ProgressBarTheme.Clone();
-
-                case DrawingSurface c:
-                    return DrawingSurfaceTheme.Clone();
-
-                case Button c:
-                    return ButtonTheme.Clone();
-
-                case Label c:
-                    return LabelTheme.Clone();
-
-                default:
-                    throw new System.Exception("Control does not have an associated theme.");
-            }
+            throw new System.Exception("Control does not have an associated theme.");
         }
 
         /// <summary>
-        /// Refreshes the theme colors of every control. Called when the <see cref="Colors"/> property has changed.
+        /// Sets a control theme based on the control type.
         /// </summary>
-        protected virtual void OnColorsChanged()
+        /// <param name="control">The control type to register a theme.</param>
+        /// <param name="theme">The theme to associate with the control.</param>
+        /// <returns>A theme that is associated with the control.</returns>
+        public void SetControlTheme(Type control, ThemeBase theme)
         {
-            ControlsConsoleTheme.RefreshTheme(Colors);
-            WindowTheme.RefreshTheme(Colors);
-
-            ScrollBarTheme.RefreshTheme(Colors);
-            ButtonTheme.RefreshTheme(Colors);
-            CheckBoxTheme.RefreshTheme(Colors);
-            ListBoxTheme.RefreshTheme(Colors);
-            ProgressBarTheme.RefreshTheme(Colors);
-            RadioButtonTheme.RefreshTheme(Colors);
-            TextBoxTheme.RefreshTheme(Colors);
-            SelectionButtonTheme.RefreshTheme(Colors);
-            DrawingSurfaceTheme.RefreshTheme(Colors);
-            LabelTheme.RefreshTheme(Colors);
+            if (null == control) throw new ArgumentNullException(nameof(control), "Cannot use a null control type");
+            _controlThemes[control] = theme ?? throw new ArgumentNullException(nameof(theme), "Cannot set the theme of a control to null");
         }
 
         /// <summary>
-        /// Clonse this library.
+        /// Clones this library.
         /// </summary>
         /// <returns>A new instance of a library.</returns>
-        public virtual Library Clone() => new Library()
+        public Library Clone()
         {
-            Colors = Colors.Clone(),
+            var library = new Library();
 
-            ButtonTheme = (ButtonTheme)ButtonTheme.Clone(),
-            SelectionButtonTheme = (ButtonTheme)SelectionButtonTheme.Clone(),
-            ScrollBarTheme = (ScrollBarTheme)ScrollBarTheme.Clone(),
-            RadioButtonTheme = (RadioButtonTheme)RadioButtonTheme.Clone(),
-            ListBoxTheme = (ListBoxTheme)ListBoxTheme.Clone(),
-            CheckBoxTheme = (CheckBoxTheme)CheckBoxTheme.Clone(),
-            TextBoxTheme = (TextBoxTheme)TextBoxTheme.Clone(),
-            ProgressBarTheme = (ProgressBarTheme)ProgressBarTheme.Clone(),
-            DrawingSurfaceTheme = (DrawingSurfaceTheme)DrawingSurfaceTheme.Clone(),
-            LabelTheme = (LabelTheme)LabelTheme.Clone(),
+            foreach (var item in _controlThemes)
+                library.SetControlTheme(item.Key, item.Value);
 
-            ControlsConsoleTheme = ControlsConsoleTheme.Clone(),
-            WindowTheme = WindowTheme.Clone(),
-        };
+            library._colors = _colors.Clone();
+
+            return library;
+        }
     }
 }