+++ ChangeLog.md
@@ -1,3 +1,22 @@
+## 11/26/2019 V8.99.0
+
+- **Breaking changes to themes and controls**
+
+  The control themes system is now a lot simpler.
+
+  - Themes only come from two places, the control directly or from the library.
+  - Each control has a `ThemeColors` property that controls the colors of the theme. If set to `null` the colors are pulled from the parent console/window.
+  - The Console/Window types have a `ThemeColors` property that specify the colors for all controls. If set to `null` the colors are instead pulled from the library.
+  - The library has been simplified. Each control type is registered with a theme via the `Library.SetControlTheme`. To overwrite a default theme, set it in the library.
+
+  When you create a new a control, and thus have a new theme, register it with the library `SadConsole.Themes.Library.Default.SetControlTheme(typeof(yourControl), new ThemeType());` When a control is created it queries the library for a theme, which is a clone of the one registered with the library. If you want to override the library-supplied theme, just set the `Control.THeme` property.
+
+  Setting colors for a console or a control is easy. Each control and `ControlsConsole` or `Window` has a `ThemeColors` property. Set this to 
+
+I know I'm not supposed to add breaking changes without increasing the major version. However, V9 is already labeled and I wasn't planning on this change. This change was actually created for V9 as a way to solve the problems with v8's theme system. But it proved so simple to implement that I wanted to port it.
+
+Because of the breaking change, I'm setting this version to a strange one, *v8.99*.
+
 ## 11/04/2019 V8.9.1
 
 - Fix `FontMaster.Rows` calculation (Thanks Kaev)